Move partial parsing groups test to separate file
